Improve catalog generation (#546)
* Improve catalog generation

This commit improves the generated catalog file when we inline other
Micronaut version catalogs. For example, before, if the platform
included `micronaut-core-bom`, then in the platform we would inline
the entries of the core bom catalog so that the user can add a
dependency using, say:

    implementation(mn.micronaut.http.server.netty)

The problem with the previous implementation is that the version which
was associated with the `micronaut.http.server.netty` was empty:

```toml
[libraries]
...
micronaut-http-server-netty = {group = "io.micronaut", name = "micronaut-http-server-netty", version = "" }
...
```

Now, this will generate:

```toml
...
micronaut-http-server-netty = {group = "io.micronaut", name = "micronaut-http-server-netty", version.ref = "micronaut-core" }
...
```

Note that the previous version worked for users without problem, because
we automatically import the BOM files, so the versions are picked from the
BOM. However, if a user overrides the version of the `micronaut.core` property
when importing the platform catalog, then **only** the core BOM version would
be affected, which can lead to a surprising behavior, as for example one
could expect the `core` dependency to be upgraded, when in fact only the BOM
would be, leaving the `core` version unchanged.

This change is required to properly fix https://github.com/micronaut-projects/micronaut-gradle-plugin/issues/668
and should ideally be merged and released before releasing platform M1.

In addition, we have improved warnings, so that when a catalog is inlined
and that we find a duplicate alias or version alias, then the error message
tells us in which catalog it was found. For example, it now explains that
the log4j dependencies are duplicates because they are defined both in
the core BOM and the logging BOM:

```
[Warning] While inlining micronaut-logging-bom-1.0.0-M1.toml, version alias 'slf4j-nop' is already defined in the catalog by [micronaut-core-bom-4.0.0-M1.toml] so it won't be imported
[Warning] While inlining micronaut-logging-bom-1.0.0-M1.toml, alias 'slf4j-simple' is already defined in the catalog by [micronaut-core-bom-4.0.0-M1.toml] so it won't be imported
[Warning] While inlining micronaut-logging-bom-1.0.0-M1.toml, alias 'logback-classic' is already defined in the catalog by [micronaut-core-bom-4.0.0-M1.toml] so it won't be imported
[Warning] While inlining micronaut-logging-bom-1.0.0-M1.toml, version alias 'slf4j-jul-to-slf4j' is already defined in the catalog by [micronaut-core-bom-4.0.0-M1.toml, micronaut-logging-bom-1.0.0-M1.toml] so it won't be imported
[Warning] While inlining micronaut-logging-bom-1.0.0-M1.toml, version alias 'slf4j-log4j12' is already defined in the catalog by [micronaut-core-bom-4.0.0-M1.toml, micronaut-logging-bom-1.0.0-M1.toml] so it won't be imported
[Warning] While inlining micronaut-logging-bom-1.0.0-M1.toml, version alias 'logback-access' is already defined in the catalog by [micronaut-core-bom-4.0.0-M1.toml] so it won't be imported
[Warning] While inlining micronaut-logging-bom-1.0.0-M1.toml, version alias 'logback-core' is already defined in the catalog by [micronaut-core-bom-4.0.0-M1.toml, micronaut-logging-bom-1.0.0-M1.toml] so it won't be imported
[Warning] While inlining micronaut-logging-bom-1.0.0-M1.toml, version alias 'slf4j-osgi-over-slf4j' is already defined in the catalog by [micronaut-core-bom-4.0.0-M1.toml, micronaut-logging-bom-1.0.0-M1.toml] so it won't be imported
[Warning] While inlining micronaut-logging-bom-1.0.0-M1.toml, alias 'slf4j-api' is already defined in the catalog by [micronaut-core-bom-4.0.0-M1.toml] so it won't be imported
[Warning] While inlining micronaut-logging-bom-1.0.0-M1.toml, version alias 'slf4j-reload4j' is already defined in the catalog by [micronaut-core-bom-4.0.0-M1.toml, micronaut-logging-bom-1.0.0-M1.toml] so it won't be imported
[Warning] While inlining micronaut-logging-bom-1.0.0-M1.toml, version alias 'slf4j-ext' is already defined in the catalog by [micronaut-core-bom-4.0.0-M1.toml, micronaut-logging-bom-1.0.0-M1.toml] so it won't be imported
[Warning] While inlining micronaut-logging-bom-1.0.0-M1.toml, version alias 'slf4j-jcl-over-slf4j' is already defined in the catalog by [micronaut-core-bom-4.0.0-M1.toml, micronaut-logging-bom-1.0.0-M1.toml] so it won't be imported
[Warning] While inlining micronaut-logging-bom-1.0.0-M1.toml, version alias 'slf4j-log4j-over-slf4j' is already defined in the catalog by [micronaut-core-bom-4.0.0-M1.toml, micronaut-logging-bom-1.0.0-M1.toml] so it won't be imported
[Warning] While inlining micronaut-openapi-bom-5.0.0-M1.toml, alias 'slf4j-nop' is already defined in the catalog by [micronaut-logging-bom-1.0.0-M1.toml] so it won't be imported
```

* Fix tests