Make it possible to read logback version from catalog
This commit adds the ability to read the default logback version
from the `mnLogging` catalog instead of the `libs` catalog or
Gradle properties. It also removes the default version as it is
likely to be outdated in any case and raise problems.

The same should probably done for objenesis and bytebuddy but
it's unclear which projects use it in practice.

Fixes #561