Remove default version for objenesis/bytebuddy
In addition, dependencies will only be added if the version is found in
a catalog (this is also true for logback version).

For bytebuddy and objenesis, the version will be fetched from `libs`,
but also from `mnTest`. Currently, the Micronaut Test catalog doesn't
declare these, but it feels like "the right place".