Make it possible to read logback version from catalog (#562)
* Make it possible to read logback version from catalog

This commit adds the ability to read the default logback version
from the `mnLogging` catalog instead of the `libs` catalog or
Gradle properties. It also removes the default version as it is
likely to be outdated in any case and raise problems.

The same should probably done for objenesis and bytebuddy but
it's unclear which projects use it in practice.

Fixes #561

* Bump JDK used to build the plugins

* Remove default version for objenesis/bytebuddy

In addition, dependencies will only be added if the version is found in
a catalog (this is also true for logback version).

For bytebuddy and objenesis, the version will be fetched from `libs`,
but also from `mnTest`. Currently, the Micronaut Test catalog doesn't
declare these, but it feels like "the right place".

* Automatically import the `mnLogging` catalog

* Bump plugin version to 6.5.x