Workaround for ClassCastException (#587)
The Sonar plugin (rightfully) assumes that options added to
the compiler are instances of `String` because typed as a
`List<String>`. Whenever compilerArgs.stream() is called,
a class cast exception happens. However, build scripts often
add arguments as GString in Groovy scripts.

To workaround this, we configure the compiler options by
coercing to a `String` in an `afterEvaluate`.