Provide option to override catalog versions from CLI
This commit adds the ability to override the versions of dependencies
declared in the catalogs that the build uses. For example, to override
the `libs.managed.jackson` version, you can specify on the command-line:

`./gradlew :foo -Poverride.libs.managed.jackson=2.6.0`

Alternatively, it is possible to provide an override properties file:

`./gradlew foo -Poverride.file.libs=overrides.properties`

and the properties file must contain the following entry:

`managed.jackson=2.6.0`

in order to behave the same as the 1st example.