Fix BOMs not being inlined
This commit fixes a problem where some aliases were not inlined
in the platform BOM. For example, the `junit` BOM dependency
is missing from the platform, despite `micronaut-test` including
it.

The reason is that in `micronaut-test`, `junit` is a managed
version, but for a BOM (alias `boms-...`). There were 2 ways
to fix this problem:

1. include all versions from the embedded catalog during
inlining. This worked, but with a drawback: only the version
would be included (e.g `junit`), but no library would be present.
Which means that the user would still have to figure out the
coordinates of, say, the junit BOM.
2. the fix in this PR, which is to include `boms-` dependencies
into the generated catalog. This works well and should be
backwards compatible, since we're only adding new entries to
the catalogs.

It's worth noting that we _don't_ do this for `boms-micronaut-`
entries, which, in this case, are available as independent
version catalogs. Including them would lead to potentially
many BOM inclusion conflicts, since many projects may include
other Micronaut BOMs.

Fixes #611