Improve binary compatiblity activation (#632)
Often when a new module is added to a project, we need to disable
binary compatibility checks for that module, because there's no
previous release yet. To do this, developers often chose the easy
way, which is to set `enabled = false` to the binary compatibility
extension.

However, it is easy to forget to re-enable binary compatibility
checks after a release. This PR adds a new method to the binary
compatibility extension to express the fact that binary checks
should be enabled once release X is done.

For example, instead of writing:

```java
micronautBuild {
    binaryCompatibility {
        String v = projectVersion - '-SNAPSHOT'
        def (int major, int minor, int patch) = (v.split(/[.]/) as List).collect { Integer.parseInt(it) }
        enabled = major > 2 || (major == 2 && minor > 2) || (major == 2 && minor == 2 && patch > 0)
    }
}
```

one can now simply write:

```java
micronautBuild {
    binaryCompatibility {
        enabledAfter '2.2.0'
    }
}
```