Inline dependencies from Micronaut BOMs into the generated BOM
This commit adds inlining of dependencies from imported Micronaut BOMs
into the generated BOM. The dependencies are added _before_ the imported
BOMs so that we make sure that they take precedence in Maven.

In addition, we still have a <scope>import</scope> dependency on the
Micronaut BOMs because they themselves may import other BOMs which are
not Micronaut BOMs. If we removed the import then we would loose some
dependency constraints.

Fixes #689