Replaced detached configurations with regular configurations
Use of detached configurations is fine if you want it to be an internal
implementation details. However, we have a use case where it's needed
to perform dependency substitution on the inlined BOMs. Unfortunately
such rules are not available from consumers when using detached configurations.

Since here we have a fixed number of configurations, we can turn them into
regular configurations instead.