Add option to select catalog to use as source for the BOM
This commit introduces a new "catalogName" property to the BOM
extension, which defaults to "libs", and tells the BOM plugin
which catalog to use for the BOM generation. This is required
for the use case where multiple BOMs have to be generated, but
using a different TOML file as source.