Fix "Duplicate entry for alias" error
In case of nested BOM inlining, it was possible that we missed a case
where an alias was already included, which lead to the _Gradle_ (not ours)
warning "Duplicate entry for alias". However, we should avoid these,
because we already capture duplicate entries (which may be fine in case
we are actually overriding nested catalog versions).