Fix duplicate entries in generated BOM or catalog (#725)
* Fix potential duplicate entries in BOM/catalog

If a dependency was managed and that it also was part of the
list of modules for which version inference should be triggered,
then we would end up with duplicate entries in the BOM.

This commit fixes that by checking, when version inference occurs,
if a library with the same GAV coordinates was already registered.

* Fix "Duplicate entry for alias" error

In case of nested BOM inlining, it was possible that we missed a case
where an alias was already included, which lead to the _Gradle_ (not ours)
warning "Duplicate entry for alias". However, we should avoid these,
because we already capture duplicate entries (which may be fine in case
we are actually overriding nested catalog versions).