Support publishing of an included build (#738)
This commit adds support for a special use case, which is currently
only used in the Micronaut Fuzzing project, which is to publish a
module which is an included build of a Micronaut repository. This
module is also a Gradle plugin, which requires special setup.
Therefore, this commit only makes sure that this use case is working.

It also fixes an issue with the "useDevelopmentVersion" code which
failed if the `checkouts` directory wasn't created in advance.