Add ossIndexAudit to the build plugins (#775)
This will automatically configure the ossIndex audit task to all Micronaut projects.
Today this relies on adding manual configuration in each project, which causes
errors like the one seen on Micronaut Test Resources, which is not configured.
Instead of duplicating code in each project, we should put this in Micronaut Build
instead.