Use Publisher API for publishing releases (#774)
This commit introduces a task to publish a release using Maven Central's
Publisher API instead of the good old Gradle publish task. The rationale
for this change is that since we moved to the new Central publisher, we
have builds which sometimes miss artifacts.

The Publisher API makes it possible to publish everything as a unit,
making sure that all artifacts are uploaded.

It's worth noting that the Publisher API doesn't support publishing
SNAPSHOTs, we have to stay with the legacy publishing tasks for this.

Using this will require a couple changes (despite upgrading the build
plugins) :

- the root project must apply the `io.micronaut.build.internal.parent`
plugin
- the release workflow need to be updated to use the new task

The first one has to be changed on every project, so the upgrade is
not immediate. The second one has to be done on the project template.
Note however that upgrading the template will trigger PRs to all
projects, which would not have the new plugin in their root project,
so it has the potential to break, hence the new minor upgrade.

Publishing from a workflow only requires calling:

    ./gradlew publishToMavenCentral --publishing-type=AUTOMATIC

The AUTOMATIC mode will automatically promote to Maven Central if
validation passes. Otherwise, we would use the manual mode, which
requires logging in the UI to promote.