Fix binary compatibility checks
In previous releases of Gradle (8.x) we were using the root project
configurations container to create detached configurations, so that
when we resolve a previous version of the current project, it would
download it like any other external dependency. This worked but
trigerred a deprecation warning for Gradle 9 saying that the
resolution was done from another context, which would break in 9.

Unfortunately, there is no way to fix this, apparently, using
public APIs. Therefore, I remembered that years ago I had introduced
an API to get truly isolated resolvers. This API was hidden, so
now it's internal, but it does exactly what we need.

This fixes the error we see for example at https://github.com/micronaut-projects/micronaut-test-resources/actions/runs/17290579530/job/49076754922?pr=900#step:8:5687

which is:

> Failed to calculate the value of task ':micronaut-test-resources-bom:checkVersionCatalogCompatibility' property 'baseline'.
   > Could not resolve all files for configuration ':micronaut-test-resources-bom:detachedConfiguration1'.
      > Could not resolve io.micronaut.testresources:micronaut-test-resources-bom:2.9.0.
        Required by:
            project :micronaut-test-resources-bom
         > Unable to find a matching variant of project :micronaut-test-resources-bom:
             - No variants exist

Ideally, that API should be public, it can break anytime, but I have no better
idea.