Remove toolchains (#809)
This commit is a follow-up to https://github.com/micronaut-projects/micronaut-project-template/pull/608.
By removing the system JDKs, we're highlighting another problem with our builds.
When we want to test with Java 21, we were (rightfully) trying to compile sources
with Java 17, which isn't available anymore. So this commit removes the use of
Gradle's toolchains by default, and replaces it with the good old "source/target"
compatibility.

A flag is still available in case a particular project wants to override this.
This change will make it so that now we'll be able to compile a project with
Java 17, while there's no Java 17 JDK on the machine.

However, this reintroduces a problem that we won't be able to test future Java
versions which are not supported by Gradle.

There's no ideal solution for this problem:

- if we use toolchains, then it means that we have to install multiple JDKs
on the machine: one to run Gradle, one to build the project with the expected
Java version (e.g 17) and one to run the tests (e.g Java 21). However, our
build plugins do NOT make the difference between the version used for compilation
and the one used for tests. In addition, some projects use the public Micronaut
Gradle plugin as a dependency, in which case the test resources service is
started using the default machine JDK, which causes other errors
- if we don't use toolchains (this commit), then we fix the inconsistencies
between several versions of Java of a project, we simplify the setup (single
Java version on the machine) BUT we're stuck to whatever Gradle supports

As a workaround for the 2d issue, this commit configures the `useToolchains`
flag so that if an environment variable is set, then toolchains will be used,
which will give us the opportunity to test future releases of the JDK by
installing 2 versions in a workflow. However, this may still lead to
inconsistencies between the version used for compilation and for tests,
depending on the project setup.