Remove some automatic dependencies
This commit removes some automatic dependencies which were added
to the test compile classpath:

- `groovy-test`, which is probably a dependency from the era where
tests were using plain Groovy (no Spock, no JUnit). This was causing
JUnit 4 to be on classpath, and some tests accidentally used wrong
annotations and/or assertions, which was causing the tests to be
ignored.
- `byte-buddy`
- `objenesis`

The last 2 being infrequently (if ever) used.