Introduce new base Java and Kotlin plugins
This commit introduces a couple new plugins, which can be used for
"test suites" or "doc examples" in Micronaut builds. Before this
change, there wasn't any base plugin for these kind of projects.
This means that everytime we update the Micronaut Build plugins
to change the baseline, for example to Java 21 in Micronaut 5,
then all build scripts of test suites or examples have to be
aligned.

The 2 new plugins are:

- `io.micronaut.build.internal.java-base`, which can be applied
to Java projects (which are NOT Micronaut modules)
- `io.micronaut.build.internal.kotlin-base`, which can be applied
to Kotlin projects (in which case the Kotlin plugin needs to be
applied on that project too)