Add default versions (#814)
This commit adds a number of default versions for `micronaut-docs`,
`micronaut-logging`, `micronaut-test`, `groovy`, `spock`, `junit5`,
`junit6` and `logback`.

This makes it redundant to define them in projects, but will
require more frequent upgrades of Micronaut Build in case we want
to force bump all versions.

There is a new choice for JUnit which is to use JUnit 6.

There is NO default version for Kotlin yet, because the version
is not a simple dependency: it's a build plugin.