Refactor build
This commit refactors the build from a mono project setup to a
multi-project setup. This is done because we now have use cases
where we'd not want some plugins on the build classpath (e.g the
Kotlin plugin), which causes some troubles with Gradle classloading.

By using a multi-project setup, we have more opportunities to tell
what are the dependencies of each project, leaving to the user the
choice to add it to classpath.

In the process the build was modernized using convention plugins
and migrated to Kotlin DSL.