Introduce Kotlin plugin
This commit introduces a new Kotlin plugin. Compared to the base
Kotlin plugin, this version introduces a transitive dependency to
the Kotlin plugin, therefore ships with a preferred Kotlin version.

In order to do so, the plugin had to be isolated from the main
build plugins, so that it can be applied independently. However,
doing so makes the user build more complicated, because now the
build plugin versions would have to be declared twice: once in
the `settings.gradle` file, and once in the `buildSrc/build.gradle`
file. To avoid this, a convenience was added: the settings plugin
will write a `buildSrc/gradle.properties` file with the current
version of the Micronaut Build plugins. Since that code is executed
_before_ the buildSrc project is evaluated, then the build script
can reference that property, without having to hardcode the
version of the build plugins. This has the advantage that we
don't have to duplicate the plugins versions.

There are 3 Kotlin plugins which can be applied:

- `[...].kotlin` : plain Kotlin support
- `[...].kotlin-kapt` : Kotlin + KAPT support
- `[...].kotlin-ksp` : Kotlin + KSP support

So the plugins come with their preferred versions
of KSP/KAPT.