Generate Micronaut Module Info descriptors
This is the Micronaut Build side of https://github.com/micronaut-projects/micronaut-core/pull/12192

An extension is added on `micronautBuild` which allows configuring the
descriptor. By default, it will use the group id and artifact id of
each module of the project to generate a descriptor. If a project is
named `-core`, then it is considered the parent of other modules of
a project. This is done because neither the `bom` nor the real `parent`
project are shipped with a jar which could carry the parent descriptor.
Therefore, we consider that the `core` module of a project is its
parent. Then all other modules are children of it.